requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
Command===0.1.0
Cybernator==0.6.2
discord.py==1.6.0
DiscordUtils==1.2.5
dnspython==1.16.0
idna==2.10
isort==5.6.3
lavalink==3.1.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
Pillow==7.1.2
pip==20.3.1
pycparser==2.20
pylint==2.6.0
pynacl==1.3.0
requests==2.24.0
simpledemotivators==1.1.1
six==1.15.0
soupsieve==2.0.1
spotipy==2.16.0
tabulate==0.8.7
toml==0.10.1
urllib3==1.25.10
websockets==8.1
wrapt==1.12.1
yarl==1.6.0
youtube-dl==2020.9.20
psycopg2==2.8.6
dj-database-url==0.5.0